requirements.txt
asgiref==3.5.2
boto3==1.26.16
botocore==1.29.16
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
crispy-bootstrap5==0.7
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.0
django-countdowntimer-model==0.0.7
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-ses==3.2.2
django-utils-six==2.0
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
mysqlclient==2.1.1
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2020.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==4.0.2
twilio==7.15.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
